Remove per-recipient resend buttons
