+++ res/layout/message_recipient_list_item.xml
@@ -73,22 +73,6 @@
                 android:visibility="gone"
                 tools:visibility="visible" />
 
-        <Button android:id="@+id/resend_button"
-                android:layout_width="wrap_content"
-                android:layout_height="38sp"
-                style="@style/InfoButton"
-                android:paddingLeft="10dp"
-                android:paddingRight="10dp"
-                android:paddingTop="5dp"
-                android:paddingBottom="5dp"
-                android:layout_gravity="center_vertical"
-                android:drawableLeft="@drawable/ic_refresh_white_18dp"
-                android:drawablePadding="2dp"
-                android:text="@string/message_recipients_list_item__resend"
-                android:visibility="gone"
-                tools:visibility="visible" />
-
-
     </LinearLayout>
 
 </org.thoughtcrime.securesms.MessageRecipientListItem>
